requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
langchain
langchain-core
langchain-google-genai
